refactor: Rename appRoute & remove auth routes
